Close the socket on disconnect, just don't free the SSL context.
This should not really make a difference in practice, but it is good form to
guarantee the socket is closed right away. That way it gets closed even in the
rare case that the user is not currently in Polaris_Work() or Polaris_Run() when
Polaris_Disconnect() is called _and_ they do not call Polaris_Free() to cleanup.

Note that in that rare case there would still be a memory leak of the SSL
context, so it is still important to call Polaris_Free().